add pageID ienumerable overload

+++ PdfToHtmlNet.Tests/UnitTests.cs
@@ -1,25 +1,26 @@
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using HtmlAgilityPack;
 using Xunit;
+using System.Diagnostics;
 
 
 namespace PdfToHtmlNet.Tests
 {
     public class UnitTests
     {
+        static string sourcePdf = @"..\..\..\TestData\0055-CPC-4.1.8.20.201-AK1.PL-0002_01_ER.pdf";
+        static string targetHtml = @"..\..\..\TestData\0055-CPC-4.1.8.20.201-AK1.PL-0002_01_ER.html";
         [Fact]
         public void Test_1()
         {
-            string sourcePdf = @"..\..\..\TestData\0055-CPC-4.1.8.20.201-AK1.PL-0002_01_ER.pdf";
-            string targetHtml = @"..\..\..\TestData\0055-CPC-4.1.8.20.201-AK1.PL-0002_01_ER.html";
-            int pageId = 0;
-            if (File.Exists(targetHtml))
-                File.Delete(targetHtml);
+            var pageID = new int[] { 4 };
+            if (File.Exists(targetHtml)) File.Delete(targetHtml);
             PdfToHtmlNet.Converter c = new PdfToHtmlNet.Converter();
             c.Encoding = Encoding.UTF8;
-            c.Convert(sourcePdf, targetHtml, pageId);
+            c.Convert(sourcePdf, targetHtml, pageID);
             Exception excep = null;
             try
             {
@@ -30,6 +31,8 @@ namespace PdfToHtmlNet.Tests
             catch (Exception ex)
             {
                 excep = ex;
+                if (excep.Message.Length > 0)
+                    Debug.Print(excep.Message);
             }
             Assert.True(excep is null);
         }
